Раскидал utils.js
